requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
loguru